requirements.txt
fastapi
uvicorn[standard]
pydantic
asyncpg
openfga-sdk
python-dotenv
httpx

